requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
spacy
scikit-learn
PyMuPDF
python-docx
